+++ src/filterConfig.js
@@ -4,6 +4,21 @@ import {
   itemFilterRenderer,
 } from './components';
 
+// Function which takes a filter name and returns
+// another function which can be used in filter config
+// to parse a given filter into a CQL manually.
+const parseFilter = name => values => {
+  if (values.length === 2) {
+    return `${name}="*"`;
+  } else if (values.length === 1 && values[0] === 'false') {
+    return `${name}="*" not ${name}="true"`;
+  } else {
+    const joinedValues = values.map(v => `"${v}"`).join(' or ');
+
+    return `${name}=${joinedValues}`;
+  }
+};
+
 export const instanceFilterConfig = [
   {
     name: 'effectiveLocation',
@@ -39,11 +54,13 @@ export const instanceFilterConfig = [
     name: 'staffSuppress',
     cql: 'staffSuppress',
     values: [],
+    parse: parseFilter('staffSuppress'),
   },
   {
     name: 'discoverySuppress',
     cql: 'discoverySuppress',
     values: [],
+    parse: parseFilter('discoverySuppress'),
   },
 ];
 
@@ -95,6 +112,7 @@ export const holdingFilterConfig = [
     name: 'discoverySuppress',
     cql: 'holdingsRecords.discoverySuppress',
     values: [],
+    parse: parseFilter('holdingsRecords.discoverySuppress'),
   },
 ];
 
@@ -135,6 +153,7 @@ export const itemFilterConfig = [
     name: 'discoverySuppress',
     cql: 'item.discoverySuppress',
     values: [],
+    parse: parseFilter('item.discoverySuppress'),
   }
 ];